pyproject.toml
[build-system]
requires = ["uv_build>=0.9.22,<0.10.0"]
build-backend = "uv_build"

[project]
name = "exam_project"
version = "0.0.1"
description = "Contains code for exam project for DTU course MLOps"
authors = [
  { name = "Group 29", email = "your@email.com" },
]

keywords = ["machine learning", "MLOps"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Programming Language :: Python :: 3",
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dvc[gcs]>=3.65.0",
    "dvc-gdrive>=3.0.1",
    "dvc-gs>=3.0.2",
    "fastapi==0.115.6",
    "hydra-core>=1.3.2",
    "kagglehub>=0.4.0",
    "numpy>=2.4.0",
    "pillow>=12.1.0",
    "python-dotenv>=1.2.1",
    "pytorch-lightning>=2.6.0",
    "scikit-learn>=1.8.0",
    "torch==2.6.0",
    "torchvision>=0.21.0",
    "transformers>=4.57.3",
    "typer==0.15.1",
    "uvicorn==0.34.0",
    "wandb>=0.24.0",
    "torchmetrics>=1.8.2",
    "loguru>=0.7.3",
    "matplotlib>=3.10.8",
]

[dependency-groups]
dev = [
    "coverage==7.6.9",
    "invoke==2.2.0",
    "mkdocs-material==9.4.6",
    "mkdocs==1.6.1",
    "pre-commit==4.1.0",
    "pytest==8.3.4",
    "ruff==0.1.3",
    "mkdocstrings-python==1.12.2",
]


[tool.uv.sources]
torch = [
    { index = "pytorch-cpu" },
]
torchvision = [
    { index = "pytorch-cpu" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.ruff]
line-length = 120

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

[tool.coverage.run]
omit = ["tests/*"]
source = ["src"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[project.scripts]
train = "exam_project.train:train"
evaluate = "exam_project.evaluate:evaluate"
test = "exam_project.test:test"

